+++ src/sample/GameElement.java
@@ -5,19 +5,19 @@ import java.io.Serializable;
 public abstract class GameElement implements Serializable {
 	private Point position;
 
-	public float getPosX() {
+	public double getPosX() {
 		return this.position.getX();
 	}
 
-	public void setPosX(float _x) {
+	public void setPosX(double  _x) {
 		this.position.setX(_x);
 	}
 
-	public float getPosY() {
+	public double getPosY() {
 		return this.position.getY();
 	}
 
-	public void setPosY(float _y) {
+	public void setPosY(double _y) {
 		this.position.setY(_y);
 	}
 }
