+++ src/sample/Point.java
@@ -1,22 +1,22 @@
 package sample;
 
 public class Point {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
-    public float getX(){
+    public double getX(){
         return this.x;
     } 
 
-    public float getY(){
+    public double getY(){
         return this.y;
     } 
 
-    public void setX(float _x){
+    public void setX(double _x){
         this.x = _x;
     }
 
-    public void setY(float _y){
+    public void setY(double _y){
         this.y = _y;
     }
 }
